eid-project-final/requirements.txt
requests
web3
cryptography
joblib
py-solc-x
scikit-learn
cohere
pandas